bot/requirements.txt
python-telegram-bot==13.7
geopy==2.2.0
flask_security==3.0.0
flask-babelex==0.9.4